fix: Refatorando funcionalidades de característica: Tipo de Cabelo

+++ src/modules/tipoCabelo/useCases/deleteTipoCabelo/DeleteTipoCabeloUseCase.ts
@@ -2,6 +2,7 @@ import { ITipoCabeloRepository } from '@modules/tipoCabelo/repositories/ITipoCab
 import { inject, injectable } from 'tsyringe';
 
 import { AppError } from '@shared/errors/AppError';
+import { Messages } from '@shared/messages/Messages';
 
 @injectable()
 class DeleteTipoCabeloUseCase {
@@ -11,9 +12,15 @@ class DeleteTipoCabeloUseCase {
   ) {}
 
   async execute(id: string): Promise<void> {
+    if (!id) {
+      throw new AppError(
+        `${Messages.MISSING_PARAMETERS}: ID da Característica`,
+      );
+    }
+
     const tipoCabelo = await this.tipoCabeloRepository.listById(id);
     if (!tipoCabelo) {
-      throw new AppError('Product not found', 404);
+      throw new AppError(Messages.CHARACTERISTICS_NOT_FOUND, 404);
     }
 
     await this.tipoCabeloRepository.delete(id);
